# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Stroit Backend Suite API
  version: 1.0.0
  description: |
    API surface for the Stroit project gateway and microservices.
    Paths are versioned (`/v1`) and all responses follow the `{ success, data, error }` contract.
servers:
  - url: https://dev.api.stroitproject.local
    description: Development environment
  - url: https://test.api.stroitproject.local
    description: Test environment
  - url: https://api.stroitproject.local
    description: Production environment
...
paths:
...
  /v1/orders:
    post:
      tags:
        - Orders
      summary: Create a new order for the current user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderCreate"
      responses:
        "201":
          $ref: "#/components/responses/Order"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

    get:
      tags:
        - Orders
      summary: Paginated list of the caller's orders
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: perPage
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
        - in: query
          name: sort
          schema:
            type: string
            enum:
              - createdAt
              - status
            default: createdAt
        - in: query
          name: order
          schema:
            type: string
            enum:
              - asc
              - desc
            default: desc
      responses:
        "200":
          $ref: "#/components/responses/OrderList"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/orders/{orderId}:
    get:
      tags:
        - Orders
      summary: Retrieve a single order
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          $ref: "#/components/responses/Order"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/orders/{orderId}/status:
    patch:
      tags:
        - Orders
      summary: Update the order status (admin only)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          $ref: "#/components/responses/Order"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /v1/orders/{orderId}/cancel:
    post:
      tags:
        - Orders
      summary: Cancel an existing order
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          $ref: "#/components/responses/Order"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    OrderItem:
      type: object
      required:
        - productId
        - name
        - quantity
        - unitPrice
      properties:
        productId:
          type: string
        name:
          type: string
        quantity:
          type: number
          format: float
          minimum: 1
        unitPrice:
          type: number
          format: float
          minimum: 0
    OrderCreate:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
    Order:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        status:
          type: string
          enum:
            - created
            - in_progress
            - completed
            - cancelled
        total:
          type: number
          format: float
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    OrderStatusUpdate:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum:
            - created
            - in_progress
            - completed
            - cancelled
...
    OrderList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        page:
          type: integer
        perPage:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
    ErrorPayload:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
  responses:
    Success:
      description: Successful operation
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                const: true
              data:
                type: object
...
    Order:
      description: Order data
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/responses/Success"
              - type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Order"
...
    OrderList:
      description: Paginated order list
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/responses/Success"
              - type: object
                properties:
                  data:
                    $ref: "#/components/schemas/OrderList"
    BadRequest:
      description: Validation failed
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                const: false
              error:
                $ref: "#/components/schemas/ErrorPayload"
    Unauthorized:
      description: Authentication required or invalid token
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                const: false
              error:
                $ref: "#/components/schemas/ErrorPayload"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                const: false
              error:
                $ref: "#/components/schemas/ErrorPayload"
    NotFound:
      description: Resource was not found
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                const: false
              error:
                $ref: "#/components/schemas/ErrorPayload"
...